LayoutTests/http/tests/site-isolation/page-lifecycle/pagehide.html

<!-- webkit-test-runner [ SiteIsolationEnabled=true ] -->
<!-- This test is adapted from LayoutTests/imported/w3c/web-platform-tests/html/browsers/browsing-the-web/history-traversal/pageswap/pageswap-cross-origin.sub.html -->

<!DOCTYPE html>
<title>Tests that pagehide is fired when site isolation is enabled</title>

<script src="/js-test-resources/js-test.js"></script>
<script>
description("Tests that pagehide is fired when site isolation is enabled.");
jsTestIsAsync = true;

const chan = new BroadcastChannel("channel");
chan.addEventListener("message", (e) => {
  caughtEvent = e.data;
  shouldBeEqualToString("caughtEvent.name", "pagehide");
  shouldBe("caughtEvent.params.persisted", "false");
  popup.close();
  finishJSTest();
});

popup = window.open("/site-isolation/page-lifecycle/resources/navigating-away-subpage.html?event=pagehide&param=persisted");
</script>
