LayoutTests/http/tests/site-isolation/page-lifecycle/pagehide-expected.txt

Tests that pagehide is fired when site isolation is enabled.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS caughtEvent.name is "pagehide"
PASS caughtEvent.params.persisted is false
PASS successfullyParsed is true

TEST COMPLETE
